# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoicePrompt API",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/sessions": {
      "post": {
        "summary": "Create Session",
        "operationId": "create_session_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/sessions/{session_id}/chunks/{sequence}": {
      "put": {
        "summary": "Upload Chunk",
        "operationId": "upload_chunk_v1_sessions__session_id__chunks__sequence__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "sequence",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sequence"
            }
          },
          {
            "name": "X-Content-SHA256",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{64}$",
              "title": "X-Content-Sha256"
            }
          },
          {
            "name": "X-Started-Ms",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "X-Started-Ms"
            }
          },
          {
            "name": "X-Duration-Ms",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 60000,
              "exclusiveMinimum": 0,
              "title": "X-Duration-Ms"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}/complete": {
      "post": {
        "summary": "Complete Session",
        "operationId": "complete_session_v1_sessions__session_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionComplete"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transcriptions": {
      "post": {
        "summary": "Transcribe Immediately",
        "operationId": "transcribe_immediately_v1_transcriptions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Session-Id"
            }
          },
          {
            "name": "X-Duration-Ms",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 7200000,
              "exclusiveMinimum": 0,
              "title": "X-Duration-Ms"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "default": "cs-CZ",
              "title": "X-Locale"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transcript"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}": {
      "get": {
        "summary": "Get Session",
        "operationId": "get_session_v1_sessions__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transcripts": {
      "get": {
        "summary": "List Transcripts",
        "operationId": "list_transcripts_v1_transcripts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptList"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/transcripts/{transcript_id}": {
      "get": {
        "summary": "Get Transcript",
        "operationId": "get_transcript_v1_transcripts__transcript_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transcript_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transcript Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transcript"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Transcript",
        "operationId": "delete_transcript_v1_transcripts__transcript_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transcript_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transcript Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/events/token": {
      "post": {
        "summary": "Event Token",
        "operationId": "event_token_v1_events_token_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PubSubToken"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ChunkReceipt": {
        "properties": {
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "checksum": {
            "type": "string",
            "title": "Checksum"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted",
            "default": true
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "sequence",
          "checksum"
        ],
        "title": "ChunkReceipt"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PubSubToken": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "url",
          "expires_at"
        ],
        "title": "PubSubToken"
      },
      "SessionComplete": {
        "properties": {
          "expected_segment_count": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Expected Segment Count"
          }
        },
        "type": "object",
        "required": [
          "expected_segment_count"
        ],
        "title": "SessionComplete"
      },
      "SessionCreate": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "audio_format": {
            "type": "string",
            "pattern": "^(m4a|aac)$",
            "title": "Audio Format"
          },
          "locale": {
            "type": "string",
            "maxLength": 32,
            "title": "Locale",
            "default": "cs-CZ"
          }
        },
        "type": "object",
        "required": [
          "id",
          "audio_format"
        ],
        "title": "SessionCreate"
      },
      "SessionStatus": {
        "type": "string",
        "enum": [
          "created",
          "recording",
          "uploading",
          "transcribing",
          "refining",
          "completed",
          "failed"
        ],
        "title": "SessionStatus"
      },
      "SessionView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "accepted_segments": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Accepted Segments"
          },
          "expected_segment_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Segment Count"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "accepted_segments",
          "created_at",
          "updated_at"
        ],
        "title": "SessionView"
      },
      "Transcript": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "markdown": {
            "type": "string",
            "title": "Markdown"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "created_at",
          "expires_at",
          "markdown"
        ],
        "title": "Transcript"
      },
      "TranscriptList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TranscriptSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "TranscriptList"
      },
      "TranscriptSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "created_at",
          "expires_at"
        ],
        "title": "TranscriptSummary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}